web/pyproject.toml
[tool.poetry]
name = "turk"
version = "0.1.0"
description = ""
authors = ["urm8 <arctgisnil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.1"
lxml = "^4.8.0"
aiohttp = "^3.8.1"
requests = "^2.27.1"
click = "^8.0.4"
googletrans = "^3.0.0"
Django = "^4.0.3"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
translate = "^3.6.1"
django-ninja = "^0.17.0"
celery = "^5.2.6"
social-auth-app-django = "^5.0.0"
aiogram = "^2.20"
dj-database-url = "^0.5.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
grpcio-tools = {url = "https://files.pythonhosted.org/packages/53/3e/bdb69af20f03ce1ad54a65625302b137d3a040958f214cee5efa3ca0b0c4/grpcio-tools-1.44.0.tar.gz"}


[tool.aerich]
tortoise_orm = "turk.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.alembic]
